requirements.txt
pdfplumber>=0.7.0
PyMuPDF>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
pytesseract>=0.3.8
numpy>=1.19.0
transformers>=4.0.0
torch>=1.8.0
PyYAML>=5.4.0 